Removed a bug.
The bug replaced the first letter of second-last row with the separator if the string ended with two separators. The problematic code portion, which allocated the row value of second-last element to the last element, was commented out. The code is unnecessary in all other cases since the row value of last element will be equal to row value of second-last element in all cases where the string does not end with more than one separators, i.e. the second-last element is not same as last element of the string.